Refactor: modal관련 custom hook 로직 분리
